+++ src/com/ViktorVano/SpeechRecognitionAI/Audio/AudioCapture.java
@@ -7,7 +7,7 @@ import java.io.*;
 import javax.sound.sampled.*;
 
 import static com.ViktorVano.SpeechRecognitionAI.GUI.SpeechRecognitionAI.customPrompt;
-import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.BooleanFile.saveBooleanToFile;
+import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.Files.BooleanFile.saveBooleanToFile;
 import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.Variables.*;
 
 public class AudioCapture {
