+++ src/com/ViktorVano/SpeechRecognitionAI/GUI/SpeechRecognitionAI.java
@@ -7,6 +7,15 @@ import com.ViktorVano.SpeechRecognitionAI.Audio.RecordedAudio;
 import com.ViktorVano.SpeechRecognitionAI.FFNN.NeuralNetworkThread;
 import com.ViktorVano.SpeechRecognitionAI.FFNN.TrainingThread;
 import com.ViktorVano.SpeechRecognitionAI.Miscellaneous.*;
+import com.ViktorVano.SpeechRecognitionAI.Tables.Commands.WordCommand;
+import com.ViktorVano.SpeechRecognitionAI.Tables.Commands.WordCommandRouter;
+import com.ViktorVano.SpeechRecognitionAI.Tables.Responses.WordResponse;
+import com.ViktorVano.SpeechRecognitionAI.Tables.Routing.WordRouter;
+import com.ViktorVano.SpeechRecognitionAI.Tables.Routing.WordRouting;
+import com.ViktorVano.SpeechRecognitionAI.Tables.ShellCommands.ShellCommand;
+import com.ViktorVano.SpeechRecognitionAI.Tables.ShellCommands.ShellCommander;
+import com.ViktorVano.SpeechRecognitionAI.Tables.Webhooks.Webhook;
+import com.ViktorVano.SpeechRecognitionAI.Tables.Webhooks.WebhookRouter;
 import com.sun.istack.internal.NotNull;
 import javafx.animation.KeyFrame;
 import javafx.animation.Timeline;
@@ -32,19 +41,19 @@ import javafx.util.Duration;
 
 import java.util.*;
 
-import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.BooleanFile.*;
-import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.FloatFile.*;
+import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.Files.BooleanFile.*;
+import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.Files.FloatFile.*;
 import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.General.*;
-import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.IntegerFile.*;
-import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.StringFile.*;
+import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.Files.IntegerFile.*;
+import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.Files.StringFile.*;
 import static com.ViktorVano.SpeechRecognitionAI.Audio.AudioDatabase.*;
 import static com.ViktorVano.SpeechRecognitionAI.FFNN.TopologyFile.*;
 import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.Variables.*;
-import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.WebhooksFile.loadWebhooks;
-import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.WordCommandsFile.loadWordCommands;
-import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.WordResponsesFile.loadWordResponses;
-import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.ShellCommandsFile.loadShellCommands;
-import static com.ViktorVano.SpeechRecognitionAI.Miscellaneous.WordRoutingFile.*;
+import static com.ViktorVano.SpeechRecognitionAI.Tables.Webhooks.WebhooksFile.loadWebhooks;
+import static com.ViktorVano.SpeechRecognitionAI.Tables.Commands.WordCommandsFile.loadWordCommands;
+import static com.ViktorVano.SpeechRecognitionAI.Tables.Responses.WordResponsesFile.loadWordResponses;
+import static com.ViktorVano.SpeechRecognitionAI.Tables.ShellCommands.ShellCommandsFile.loadShellCommands;
+import static com.ViktorVano.SpeechRecognitionAI.Tables.Routing.WordRoutingFile.*;
 
 
 public class SpeechRecognitionAI extends Application {
