+++ src/com/ViktorVano/SpeechRecognitionAI/Miscellaneous/Variables.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 
 public class Variables {
 
-    public static int versionNumber = 20230407;
+    public static int versionNumber = 20230408;
 
     public static int recorderThreshold = 500;
     public static int wordThreshold = 500;
