+++ src/com/ViktorVano/SpeechRecognitionAI/Tables/Commands/WordCommand.java
@@ -1,4 +1,4 @@
-package com.ViktorVano.SpeechRecognitionAI.Miscellaneous;
+package com.ViktorVano.SpeechRecognitionAI.Tables.Commands;
 
 import java.io.Serializable;
