+++ src/com/ViktorVano/SpeechRecognitionAI/Tables/Commands/WordCommandRouter.java
@@ -1,4 +1,4 @@
-package com.ViktorVano.SpeechRecognitionAI.Miscellaneous;
+package com.ViktorVano.SpeechRecognitionAI.Tables.Commands;
 
 import javafx.collections.ObservableList;
