+++ src/com/ViktorVano/SpeechRecognitionAI/Tables/Responses/WordResponse.java
@@ -1,4 +1,4 @@
-package com.ViktorVano.SpeechRecognitionAI.Miscellaneous;
+package com.ViktorVano.SpeechRecognitionAI.Tables.Responses;
 
 import java.io.Serializable;
