+++ src/com/ViktorVano/SpeechRecognitionAI/Tables/Routing/WordRouter.java
@@ -1,4 +1,4 @@
-package com.ViktorVano.SpeechRecognitionAI.Miscellaneous;
+package com.ViktorVano.SpeechRecognitionAI.Tables.Routing;
 
 import javafx.collections.ObservableList;
