+++ src/com/ViktorVano/SpeechRecognitionAI/Tables/ShellCommands/ShellCommand.java
@@ -1,4 +1,4 @@
-package com.ViktorVano.SpeechRecognitionAI.Miscellaneous;
+package com.ViktorVano.SpeechRecognitionAI.Tables.ShellCommands;
 
 import java.io.Serializable;
