+++ src/com/ViktorVano/SpeechRecognitionAI/Tables/ShellCommands/ShellCommander.java
@@ -1,4 +1,4 @@
-package com.ViktorVano.SpeechRecognitionAI.Miscellaneous;
+package com.ViktorVano.SpeechRecognitionAI.Tables.ShellCommands;
 
 import javafx.collections.ObservableList;
