+++ src/com/ViktorVano/SpeechRecognitionAI/Tables/Webhooks/Webhook.java
@@ -1,4 +1,4 @@
-package com.ViktorVano.SpeechRecognitionAI.Miscellaneous;
+package com.ViktorVano.SpeechRecognitionAI.Tables.Webhooks;
 
 import java.io.Serializable;
